+++ CHANGELOG.md
@@ -10,7 +10,9 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 ### Changed
 
 - Add formula for sums of powers in Stirling numbers
+- Mathematica for sums of powers in Stirling numbers
 - Add formula for sums of powers in Eulerian numbers
+- Mathematica for sums of powers in Eulerian numbers
 - Update conclusions
 - Add formula for ordinary sums of powers